Move functions from utils to object-utils
