+++ packages/atomic/src/components/common/product-template/stencil-product-template-common.tsx
@@ -1,5 +1,5 @@
-import {aggregate} from '@/src/utils/utils';
 import {isElementNode, isVisualNode} from '@/src/utils/dom-utils';
+import {aggregate} from '@/src/utils/object-utils';
 import {
   ProductTemplate,
   ProductTemplateCondition,
